fix: Some LGVDatabase methods become optional

+++ LoggingViewKit/LoggingViewKit/Sources/LGVLoggingViewService.m
@@ -55,11 +55,18 @@ static LGVLoggingViewService *_loggingViewService = nil;
 }
 
 - (NSArray<LGVLog *> *)allLogs {
-    return [self.defaultDatabase allLogs];
+    if ([self.defaultDatabase respondsToSelector:@selector(allLogs)]) {
+        return [self.defaultDatabase allLogs];
+    }
+    else {
+        return @[];
+    }
 }
 
 - (void)deleteAllLogs {
-    [self.defaultDatabase deleteAllLogs];
+    if ([self.defaultDatabase respondsToSelector:@selector(deleteAllLogs)]) {
+        [self.defaultDatabase deleteAllLogs];
+    }
 }
 
 - (void)loggingView:(id <LGVLogging>)loggingView
@@ -142,7 +149,12 @@ static LGVLoggingViewService *_loggingViewService = nil;
                 error = [LGVError errorWithMessage:@"Failed to add the log into the database"];
             }
 
-            LGVLog *savedLog = [self.defaultDatabase logByKey:log.key];
+            LGVLog *savedLog = log;
+
+            // Recreates the object with last inserted ID like the SQLite.
+            if ([self.defaultDatabase respondsToSelector:@selector(logByKey:)]) {
+                savedLog = [self.defaultDatabase logByKey:log.key];
+            }
 
             if ([self.delegate respondsToSelector:@selector(loggingViewService:didSaveLog:ofView:withEvent:error:)]) {
                 [self.delegate loggingViewService:self didSaveLog:savedLog ofView:loggingView withEvent:event error:error];
